+++ esbuild.config.mjs
@@ -119,6 +119,7 @@ async function buildBackground() {
     readFile("modules/security-action-handler.js"),
     readFile("modules/diagnostics-action-handler.js"),
     readFile("modules/script-action-handler.js"),
+    readFile("modules/data-action-handler.js"),
     readFile("modules/message-router.js"),
     readFile("modules/gm-audio-handler.js"),
     readFile("modules/gm-menu-handler.js"),