+++ tests/message-response-map.test.js
@@ -18,6 +18,7 @@ import { SETTINGS_BACKGROUND_ACTIONS } from "../src/background/settings-action-h
 import { SECURITY_BACKGROUND_ACTIONS } from "../src/background/security-action-handler.ts";
 import { DIAGNOSTICS_BACKGROUND_ACTIONS } from "../src/background/diagnostics-action-handler.ts";
 import { SCRIPT_BACKGROUND_ACTIONS } from "../src/background/script-action-handler.ts";
+import { DATA_BACKGROUND_ACTIONS } from "../src/background/data-action-handler.ts";
 import { GM_AUDIO_ACTIONS } from "../src/background/gm-audio-handler.ts";
 import { GM_COOKIE_ACTIONS } from "../src/background/gm-cookie-handler.ts";
 import { GM_MENU_ACTIONS } from "../src/background/gm-menu-handler.ts";
@@ -39,6 +40,7 @@ const ROUTED_BACKGROUND_ACTIONS = [
   ...SECURITY_BACKGROUND_ACTIONS,
   ...DIAGNOSTICS_BACKGROUND_ACTIONS,
   ...SCRIPT_BACKGROUND_ACTIONS,
+  ...DATA_BACKGROUND_ACTIONS,
   ...GM_AUDIO_ACTIONS,
   ...GM_COOKIE_ACTIONS,
   ...GM_MENU_ACTIONS,